Move page views urls to separate file
